requirements.txt
fastapi
uvicorn
requests
numpy
loguru
python-bitget
python-dotenv
